// frontend/src/app/modules/client/core/model/student/student-detail.model.ts
export class StudentDetailModel {
  first_name: string;
  last_name: string;
  registration_id: string;
  level: string;
  roll_number: string;
  nationality: string;
  blood_group: string;
  transport: string;

  father_name: string;
  mother_name: string;
  father_occupation: string;
  mother_occupation: string;
  father_contact: string;
  mother_contact: string;
  guardian_name: string;
  guardian_contact: string;

  address: string;
  date_of_birth_ad: string;
  date_of_birth_bs: string;
  gender: string;
  enrolled_date: string;
  picture: string;


}